Backend/requirements.txt
fastapi==0.70.0
pydantic==1.10.0
beanie==1.20.0
httpx==0.19.0
uvicorn==0.15.0
openai==0.27.0
python-dotenv==0.19.1
pymongo==3.12.0
timeout-decorator==0.5.0
